Transaction pool unit tests refactoring to remove duplications
